views pkg rename & log out
