+++ app/src/main/java/com/example/palfinder/views/MainActivity.kt
@@ -6,7 +6,7 @@ import android.os.Handler
 import androidx.appcompat.app.AppCompatActivity
 import com.amplifyframework.core.Amplify
 import com.example.palfinder.R
-import com.example.palfinder.views.user.LoginActivity
+import com.example.palfinder.views.auth.LoginActivity
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
