+++ app/src/main/java/com/example/palfinder/views/auth/LoginActivity.kt
@@ -1,10 +1,8 @@
-package com.example.palfinder.views.user
+package com.example.palfinder.views.auth
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.example.palfinder.R
-import com.google.android.material.tabs.TabLayout
-import kotlinx.android.synthetic.main.activity_log_in.*
 
 class LoginActivity : AppCompatActivity(R.layout.activity_log_in) {
     override fun onCreate(savedInstanceState: Bundle?) {
