+++ app/src/main/res/layout/activity_log_in.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.user.LoginActivity">
+    tools:context=".views.auth.LoginActivity">
 
     <androidx.constraintlayout.utils.widget.ImageFilterView
         android:id="@+id/friends_landscape"
