+++ app/src/main/res/values/strings.xml
@@ -33,4 +33,5 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="title_confirm_sign_up">Time to validate your account!</string>
     <string name="security_code">Security Code</string>
+    <string name="log_out">Log out</string>
 </resources>
